+++ mongo/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## To be Released
 
 * build(deps): update `github.com/Scalingo/go-utils/errors` from v2 to v3
+* refactor: replace `github.com/pkg/errors` with `github.com/Scalingo/go-utils/errors/v3`
 
 ## v1.5.2